requirements.txt
torch
torchvision
pytorch_pfn_extras
numpy
matplotlib
scikit-image
pillow
streamlit
streamlit-drawable-canvas